refactor:rename saveHistory file
